refactor: operations are in a new package Operation implements Stringer
